Revert "Move all components into separate folder"
This reverts commit f1027f012c72b9cf1d81103938cf39f4decc0dfd.
