+++ src/Content.js
@@ -4,8 +4,8 @@ import Logger from "js-logger";
 import PropTypes from "prop-types";
 import React from "react";
 
-import AdminContext from "../context/AdminContext";
 import Container from "./Container";
+import AdminContext from "./context/AdminContext";
 
 const logger = Logger.get("bananas");
