+++ src/ProgressBar.js
