+++ src/index.js
@@ -1,14 +1,14 @@
-import App from "./components/Admin";
+import App from "./Admin";
 
-export { default as Container } from "./components/Container";
-export { default as Content } from "./components/Content";
-export { default as Link } from "./components/Link";
-export { default as TitleBar } from "./components/TitleBar";
-export { default as ToolBar } from "./components/ToolBar";
-export { default as Tools } from "./components/Tools";
+export { default as Container } from "./Container";
+export { default as Content } from "./Content";
 export { default as AdminContext, useAdmin } from "./context/AdminContext";
 export { default as PageContext, usePage } from "./context/PageContext";
+export { default as Link } from "./Link";
 export { default as LoginForm } from "./pages/LoginPageForm";
+export { default as TitleBar } from "./TitleBar";
+export { default as ToolBar } from "./ToolBar";
+export { default as Tools } from "./Tools";
 export { t, Translate } from "./utils";
 
 const Bananas = {
