+++ src/pages/MePage.js
@@ -3,9 +3,9 @@ import { withStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import React from "react";
 
-import Content from "../components/Content";
-import TitleBar from "../components/TitleBar";
+import Content from "../Content";
 import AdminContext from "../context/AdminContext";
+import TitleBar from "../TitleBar";
 import ChangePasswordForm from "./ChangePasswordForm";
 import SettingsForm from "./SettingsForm";
